Also extract schema version when reading definitions back
