+++ src/DatamodelRegistryIOHelpers.cc
@@ -1,8 +1,40 @@
 #include "podio/utilities/DatamodelRegistryIOHelpers.h"
+
 #include <algorithm>
+#include <charconv>
 #include <iterator>
 
 namespace podio {
+namespace detail {
+  podio::SchemaVersionT extractSchemaVersion(const std::string_view definition) {
+    // Extract schema_version from JSON definition without full parsing
+    // Look for "schema_version": followed by a number
+    constexpr std::string_view schemaVersionKey = "\"schema_version\":";
+    if (auto pos = definition.find(schemaVersionKey); pos != std::string_view::npos) {
+      pos += schemaVersionKey.length();
+      // Skip whitespace
+      while (pos < definition.length() && std::isspace(definition[pos])) {
+        ++pos;
+      }
+      // Extract the number
+      auto start = pos;
+      while (pos < definition.length() && std::isdigit(definition[pos])) {
+        ++pos;
+      }
+      if (pos > start) {
+        // Convert substring to integer using std::from_chars for better error handling
+        podio::SchemaVersionT schemaVersion = 0;
+        auto result = std::from_chars(definition.data() + start, definition.data() + pos, schemaVersion);
+        if (result.ec == std::errc{}) {
+          return schemaVersion;
+        }
+      }
+    }
+
+    // Return 0 if no valid schema version found
+    return 0;
+  }
+} // namespace detail
 
 void DatamodelDefinitionCollector::registerDatamodelDefinition(const podio::CollectionBase* coll,
                                                                const std::string& name) {
@@ -54,6 +86,14 @@ std::optional<podio::version::Version> DatamodelDefinitionHolder::getDatamodelVe
     return std::get<1>(*it);
   }
 
+  return std::nullopt;
+}
+
+std::optional<podio::SchemaVersionT> DatamodelDefinitionHolder::getSchemaVersion(const std::string& name) const {
+  const auto definition = getDatamodelDefinition(name);
+  if (definition != "{}") {
+    return detail::extractSchemaVersion(definition);
+  }
   return std::nullopt;
 }
